requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
black==24.2.0
celery==5.3.6
cfgv==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.3
coverage-badge==1.1.0
distlib==0.3.8
Django==5.0.2
filelock==3.13.1
identify==2.5.35
kombu==5.3.5
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
prompt-toolkit==3.0.43
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
